pyproject.toml
[tool.poetry]
name = "eresumao"
version = "0.1.0"
description = ""
authors = ["ERESUMAO <eresumao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.24.2"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
